Storing data in file via the model using fs module
