--- Models/product.js
@@ -1,15 +1,0 @@
-const products = [];
-
-export default class Product {
-	constructor(value) {
-		this.title = value;
-	}
-
-	save() {
-		products.push(this);
-	}
-
-	static fetchAll() {
-		return products;
-	}
-};
